SERVER: BREAKING CHANGE for submoduled projects
The events and types in server have been re-arranged and simplified. Streamer is the direct parent of request class now, instead of using a proxy. Fields not on the proxy are marked private or protected. req and res are using private field identifiers (with the `#`). Refer to the changes in router.ts and streamer.ts for the source, and in RequestState.ts and registerRequest.ts for the implementation of those changes. As always, remember that find all references should work for almost everything in this project.